test: move milhdbk217f test ATTRIBUTES to conftest.py as fixture
