+++ tests/analyses/milhdbk217f/conftest.py
@@ -1,0 +1,64 @@
+# Third Party Imports
+import pytest
+
+
+@pytest.fixture
+def test_attributes():
+    yield {
+        "hardware_id": 12,
+        "category_id": 4,
+        "subcategory_id": 1,
+        "add_adj_factor": 0.0,
+        "application_id": 1,
+        "area": 1.5,
+        "capacitance": 0.0000033,
+        "configuration_id": 1,
+        "construction_id": 1,
+        "contact_form_id": 1,
+        "contact_gauge": 20,
+        "contact_rating_id": 1,
+        "current_operating": 0.14,
+        "current_rated": 0.5,
+        "current_ratio": 0.28,
+        "duty_cycle": 100.0,
+        "environment_active_id": 1,
+        "family_id": 1,
+        "feature_size": 1.5,
+        "frequency_operating": 1.5,
+        "hazard_rate_method_id": 1,
+        "insert_id": 1,
+        "insulation_id": 1,
+        "matching_id": 1,
+        "mult_adj_factor": 1.0,
+        "n_active_pins": 14,
+        "n_circuit_planes": 3,
+        "n_cycles": 32,
+        "n_elements": 8,
+        "n_hand_soldered": 5,
+        "n_wave_soldered": 138,
+        "package_id": 1,
+        "power_operating": 0.5,
+        "power_rated": 0.75,
+        "power_ratio": 0.67,
+        "quality_id": 2,
+        "quantity": 1,
+        "resistance": 0.05,
+        "specification_id": 1,
+        "technology_id": 1,
+        "temperature_active": 45.0,
+        "temperature_case": 38.2,
+        "temperature_rated_max": 105.0,
+        "temperature_rise": 10.0,
+        "theta_jc": 12.0,
+        "type_id": 1,
+        "voltage_ac_operating": 0.04,
+        "voltage_dc_operating": 3.3,
+        "voltage_rated": 12.0,
+        "voltage_ratio": 0.54,
+        "weight": 0.5,
+        "years_in_production": 3,
+        "hazard_rate_active": 0.0,
+        "lambda_b": 0.0,
+        "piE": 2.0,
+        "piQ": 0.0,
+    }
